+++ impressionistDoc.h
@@ -29,7 +29,7 @@ public:
 	void	setSize(int size);				// set the UI size
 	int		getLineWidth();
 	int		getLineAngle();
-	float   getAlpha();
+	double  getAlpha();
 	char*	getImageName();					// get the current image name
